Use floorf for float floor, fix README formatting

+++ runtime/runtime.c
@@ -168,7 +168,7 @@ INLINE float f32_max(float a, float b) {
 }
 
 INLINE float f32_floor(float a) {
-    return floor(a);
+    return floorf(a);
 }
 
 INLINE double f64_min(double a, double b) {
